% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_summ_count.R
\name{ff_summ_count_unique_by_groups}
\alias{ff_summ_count_unique_by_groups}
\title{By Multiple Groups, Count the Number of Unique Observations (Individuals) within Group}
\usage{
ff_summ_count_unique_by_groups(df, ar_svr_group = c("S.country",
  "vil.id"), svr_unique_identifier = "indi.id")
}
\arguments{
\item{df}{dataframe input dataframe of interest}

\item{ar_svr_group}{array string array of variables to group by}

\item{svr_unique_identifier}{string variable that has the unique key of interest}
}
\value{
a dataframe with stats outputs.
}
\description{
We have multiple groups (country, village), we want to know the number of unique observations within these groups.
In addition, we also want to generate the total number of observations for each variable within these
country/village groups, these total observations includes multiple values for each unique individual.
}
\examples{
df_uniques_count_by_vil <- ff_summ_count_unique_by_groups(df_hgt_wgt,
                         ar_svr_group=c('S.country', 'vil.id'),
                         svr_unique_identifier = 'indi.id')
print(df_uniques_count_by_vil, n=50)
df_uniques_count_by_mth <- ff_summ_count_unique_by_groups(df_hgt_wgt,
                         ar_svr_group=c('S.country', 'svymthRound'),
                         svr_unique_identifier = 'indi.id')
print(df_uniques_count_by_mth, n=50)
df_uniques_count_by_country <- ff_summ_count_unique_by_groups(df_hgt_wgt,
                         ar_svr_group=c('S.country'),
                         svr_unique_identifier = 'indi.id')
print(df_uniques_count_by_country)
}
\references{
\url{https://fanwangecon.github.io/R4Econ/reference/ff_summ_count_unique_by_groups.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
