% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_summ_percentiles.R
\name{ff_summ_percentiles}
\alias{ff_summ_percentiles}
\title{Summarize each variable in a dataset}
\usage{
ff_summ_percentiles(df, col2varname = FALSE)
}
\arguments{
\item{df}{dataframe input dataframe of interest}

\item{col2varname}{boolean if true then rotate table}
}
\value{
a dataframe with summary statistics.
}
\description{
Generate similar statistics as what is generated by distributional statistics
calculator from dynamic asset webpage's distributional codes: \url{https://fanwangecon.github.io/CodeDynaAsset/}
}
\examples{
ff_summ_percentiles(iris)
ff_summ_percentiles(iris, TRUE)
}
\references{
\url{https://fanwangecon.github.io/R4Econ/}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
