% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_summ_percentiles.R
\name{ff_summ_percentiles_by_groups}
\alias{ff_summ_percentiles_by_groups}
\title{Percentiles mean and sd for as columns, variables as rows, groups vars super-rows}
\usage{
ff_summ_percentiles_by_groups(df, vars.group.by, it.round = 4,
  col2varname = FALSE)
}
\description{
Percentiles mean and sd for as columns, variables as rows, groups vars super-rows
}
